Fix a bug where writing parquet files over 2GB would cause invalid byte counts in the footer.
While using an int32 to count bytes was fine for most smaller structures inside parquet files, it's not ok for counting the total bytes in the file.

+++ core/src/main/java/com/markosindustries/parquito/RowGroupWriter.java
@@ -182,7 +182,7 @@ public class RowGroupWriter<Row> implements AutoCloseable, Writer.DataPageAccumu
     final var bytesBeforeFooter = byteCountingStream.getBytesWritten();
     ParquetFooter.write(fileMetaData, byteCountingStream).join();
     final var footerBytes = byteCountingStream.getBytesWritten() - bytesBeforeFooter;
-    LittleEndian.writeInt(footerBytes, byteCountingStream);
+    LittleEndian.writeInt((int)footerBytes, byteCountingStream);
     byteCountingStream.write(PARQUET_UNENCRYPTED_MAGIC_BYTES);
   }
